+++ packages/dependency-extraction-webpack-plugin/lib/util.js
@@ -11,6 +11,7 @@ const BUNDLED_PACKAGES = [
 	'@wordpress/interface',
 	'@wordpress/kebab-case',
 	'@wordpress/style-runtime',
+	'@wordpress/sync',
 	'@wordpress/ui',
 	'@wordpress/undo-manager',
 	'@wordpress/views',